+++ src/services/LineItems.php
@@ -312,6 +312,11 @@ class LineItems extends Component
             ->where(['id' => $id])
             ->one();
 
+        if ($result) {
+            // Unpack the snapshot
+            $result['snapshot'] = Json::decodeIfJson($result['snapshot']);
+        }
+
         return $result ? new LineItem($result) : null;
     }